parser support filedName like 'message_id'
