+++ src/test/java/com/alibaba/json/bvt/parser/deser/SmartMatchTest.java
@@ -6,47 +6,46 @@ import com.alibaba.fastjson.JSON;
 
 import junit.framework.TestCase;
 
-
 public class SmartMatchTest extends TestCase {
-    
-    public void f_test_0 () throws Exception {
-        String text = "{\"_id\":1001}";
-        
+
+    public void f_test_0() throws Exception {
+        String text = "{\"message_id\":1001}";
+
         VO vo = JSON.parseObject(text, VO.class);
-        Assert.assertEquals(1001, vo.getId());
+        Assert.assertEquals(1001, vo.getMessageId());
     }
-    
-    public void test_vo2 () throws Exception {
-        String text = "{\"_id\":1001}";
-        
+
+    public void test_vo2() throws Exception {
+        String text = "{\"message_id\":1001}";
+
         VO2 vo = JSON.parseObject(text, VO2.class);
-        Assert.assertEquals(1001, vo.getId());
+        Assert.assertEquals(1001, vo.getMessageId());
     }
 
     private static class VO {
 
-        private int id;
+        private int messageId;
 
-        public int getId() {
-            return id;
+        public int getMessageId() {
+            return messageId;
         }
 
-        public void setId(int id) {
-            this.id = id;
+        public void setMessageId(int messageId) {
+            this.messageId = messageId;
         }
 
     }
-    
+
     public static class VO2 {
 
-        private int id;
+        private int messageId;
 
-        public int getId() {
-            return id;
+        public int getMessageId() {
+            return messageId;
         }
 
-        public void setId(int id) {
-            this.id = id;
+        public void setMessageId(int messageId) {
+            this.messageId = messageId;
         }
 
     }
